requirements.txt
coloredlogs
angr
angr-utils
pyyaml
claripy
